move_evaluator_for_concept compiles; still need to change to dependency injection, add builder, and add tests
