+++ cmake/add_subdirectories.cmake
@@ -5,6 +5,7 @@ add_subdirectory(${SOURCE_CODE_DIR}/game_board_for_concepts)
 add_subdirectory(${SOURCE_CODE_DIR}/json_utility_nlohmann)
 add_subdirectory(${SOURCE_CODE_DIR}/json_validation)
 add_subdirectory(${SOURCE_CODE_DIR}/move_evaluators)
+add_subdirectory(${SOURCE_CODE_DIR}/move_evaluators_for_concepts)
 add_subdirectory(${SOURCE_CODE_DIR}/move_calculator)
 add_subdirectory(${SOURCE_CODE_DIR}/piece_moves)
 add_subdirectory(${SOURCE_CODE_DIR}/piece_position_points)
