+++ include/builders.hpp
@@ -3,6 +3,8 @@
 #include <game_board_for_concepts.hpp>
 #include <memory>
 #include <piece_position_points_for_concepts.hpp>
+#include <zobrist_for_concepts.hpp>
+
 namespace gameboard {
 
 class GameBoardBuilder {
@@ -14,9 +16,8 @@ public:
 
 } // namespace gameboard
 
-
 namespace piecepoints {
-    class PiecePositionPointsBuilder {
+class PiecePositionPointsBuilder {
 public:
   std::shared_ptr<PiecePositionPointsForConcepts> build(
       std::string json_file = piecepoints::kICGABPOPath
@@ -32,7 +33,26 @@ public:
   std::shared_ptr<PiecePositionPointsForConcepts> build(BPOPointsSKeys &bpo_points_skeys
   );
 };
-}
+} // namespace piecepoints
 
-
-
+namespace boardstate {
+template <typename KeyType, size_t NumConfKeys>
+class ZobristCoordinatorBuilder {
+public:
+  std::shared_ptr<ZobristCoordinatorForConcepts<KeyType, NumConfKeys>> build(
+      uint32_t prng_seed
+  ) {
+    ZobristComponentForConcepts<KeyType, NumConfKeys> zobrist_component{prng_seed};
+    return std::make_shared<ZobristCoordinatorForConcepts<KeyType, NumConfKeys>>(
+        zobrist_component
+    );
+  }
+  std::shared_ptr<ZobristCoordinatorForConcepts<KeyType, NumConfKeys>> build() {
+    auto prng_seed = (uint32_t)std::random_device{}();
+    ZobristComponentForConcepts<KeyType, NumConfKeys> zobrist_component{prng_seed};
+    return std::make_shared<ZobristCoordinatorForConcepts<KeyType, NumConfKeys>>(
+        zobrist_component
+    );
+  }
+};
+} // namespace boardstate
