+++ src/core/move_evaluators_for_concepts/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(move_evaluators_for_concepts move_evaluators_for_concepts.cpp)
+
+target_link_libraries(
+    move_evaluators_for_concepts PUBLIC
+    nlohmann_json
+    utility_functs)
